+++ src/commons.rs
@@ -1,7 +1,7 @@
-	pub const WHITE: [f32; 4] = [0.3, 0.3, 1.0, 1.0];
-	pub const BLACK: [f32; 4] = [0.0, 0.0, 0.0, 1.0];
-	pub const WIDTH: f64 = 640.;
-	pub const HEIGHT: f64 = 480.;
+pub const WHITE: [f32; 4] = [0.3, 0.3, 1.0, 1.0];
+pub const BLACK: [f32; 4] = [0.0, 0.0, 0.0, 1.0];
+pub const WIDTH: f64 = 640.;
+pub const HEIGHT: f64 = 480.;
 
 pub struct Center {
     pub x: f64,
